split http_fetch.hpp into http_fetch.hpp and http_fetch.cpp
http_fetch.hpp included <curl/curl.h> and was fully inline, so every TU
that pulled it in (mostly transitively via zarr.hpp, ~88 TUs) re-parsed
curl.h. Move all curl-typed code into http_fetch.cpp and drop the curl.h
include from the header.
