+++ volume-cartographer/utils/src/http_fetch.cpp
@@ -1,0 +1,522 @@
+#include "utils/http_fetch.hpp"
+
+#if UTILS_HAS_CURL
+
+#include <curl/curl.h>
+
+#include <cstring>
+#include <cstdio>
+#include <cstdlib>
+#include <memory>
+#include <mutex>
+#include <random>
+#include <stdexcept>
+#include <thread>
+#include <algorithm>
+
+namespace utils {
+
+// ---------------------------------------------------------------------------
+// S3 URL utilities
+// ---------------------------------------------------------------------------
+
+bool is_s3_url(std::string_view url) noexcept {
+    if (url.starts_with("s3://") || url.starts_with("S3://"))
+        return true;
+    // s3+REGION:// form
+    if ((url.starts_with("s3+") || url.starts_with("S3+")) && url.find("://") != std::string_view::npos)
+        return true;
+    return false;
+}
+
+std::optional<S3Url> parse_s3_url(std::string_view url) {
+    if (!is_s3_url(url))
+        return std::nullopt;
+
+    std::string region;
+
+    // Check for s3+REGION:// form
+    auto scheme_end = url.find("://");
+    if (scheme_end == std::string_view::npos)
+        return std::nullopt;
+
+    auto scheme = url.substr(0, scheme_end);
+    auto plus = scheme.find('+');
+    if (plus != std::string_view::npos) {
+        region = std::string{scheme.substr(plus + 1)};
+    }
+
+    auto rest = url.substr(scheme_end + 3); // skip "SCHEME://"
+    auto slash = rest.find('/');
+    if (slash == std::string_view::npos)
+        return S3Url{std::string{rest}, {}, std::move(region)};
+
+    auto bucket = rest.substr(0, slash);
+    auto key = rest.substr(slash + 1);
+    return S3Url{std::string{bucket}, std::string{key}, std::move(region)};
+}
+
+std::string s3_to_https(const S3Url& parsed) {
+    std::string result = "https://";
+    result += parsed.bucket;
+    if (!parsed.region.empty()) {
+        result += ".s3.";
+        result += parsed.region;
+        result += ".amazonaws.com";
+    } else {
+        result += ".s3.amazonaws.com";
+    }
+    if (!parsed.key.empty()) {
+        result += '/';
+        result += parsed.key;
+    }
+    return result;
+}
+
+std::string s3_to_https(std::string_view s3_url) {
+    auto parsed = parse_s3_url(s3_url);
+    if (!parsed)
+        return std::string{s3_url};
+    return s3_to_https(*parsed);
+}
+
+// ---------------------------------------------------------------------------
+// curl plumbing (file-scope; not exposed in the header)
+// ---------------------------------------------------------------------------
+namespace {
+
+struct CurlDeleter {
+    void operator()(CURL* c) const noexcept { curl_easy_cleanup(c); }
+};
+using CurlHandle = std::unique_ptr<CURL, CurlDeleter>;
+
+// Global init/cleanup (reference counted)
+struct CurlGlobal {
+    CurlGlobal() {
+        if (ref_count_.fetch_add(1, std::memory_order_relaxed) == 0)
+            curl_global_init(CURL_GLOBAL_DEFAULT);
+    }
+    ~CurlGlobal() {
+        if (ref_count_.fetch_sub(1, std::memory_order_acq_rel) == 1)
+            curl_global_cleanup();
+    }
+    CurlGlobal(const CurlGlobal&) = delete;
+    CurlGlobal& operator=(const CurlGlobal&) = delete;
+
+private:
+    static inline std::atomic<int> ref_count_{0};
+};
+
+CurlHandle make_curl() {
+    static CurlGlobal global;
+    return CurlHandle{curl_easy_init()};
+}
+
+CURL* thread_handle() {
+    thread_local CurlHandle tl_handle{nullptr};
+    if (!tl_handle) {
+        tl_handle = make_curl();
+    }
+    return tl_handle.get();
+}
+
+int xferinfo_callback(void* /*clientp*/,
+                      curl_off_t, curl_off_t,
+                      curl_off_t, curl_off_t) noexcept
+{
+    // Acquire ordering pairs with the release store in abortAll() so
+    // termination is guaranteed to be observed promptly. On x86 this
+    // is free (plain load); relaxed would have worked there but is
+    // not formally guaranteed to see the abort bit in bounded time
+    // on weakly-ordered archs.
+    return HttpClient::abort_flag().load(std::memory_order_acquire) ? 1 : 0;
+}
+
+// Write callback: append to std::vector<std::byte>
+std::size_t write_callback(char* ptr, std::size_t size,
+                           std::size_t nmemb, void* userdata) noexcept {
+    auto& buf = *static_cast<std::vector<std::byte>*>(userdata);
+    auto total = size * nmemb;
+    auto* src = reinterpret_cast<const std::byte*>(ptr);
+    buf.insert(buf.end(), src, src + total);
+    return total;
+}
+
+// Header callback: capture content-type and content-length
+std::size_t header_callback(char* ptr, std::size_t size,
+                            std::size_t nmemb, void* userdata) noexcept {
+    auto& resp = *static_cast<HttpResponse*>(userdata);
+    auto total = size * nmemb;
+    std::string_view line(ptr, total);
+
+    // Strip trailing \r\n
+    while (!line.empty() && (line.back() == '\r' || line.back() == '\n'))
+        line.remove_suffix(1);
+
+    constexpr std::string_view ct_prefix = "content-type:";
+    constexpr std::string_view cl_prefix = "content-length:";
+
+    auto lower_match = [&](std::string_view prefix) -> bool {
+        if (line.size() < prefix.size()) return false;
+        for (std::size_t i = 0; i < prefix.size(); ++i) {
+            char c = line[i];
+            if (c >= 'A' && c <= 'Z') c += 32;
+            if (c != prefix[i]) return false;
+        }
+        return true;
+    };
+
+    if (lower_match(ct_prefix)) {
+        auto val = line.substr(ct_prefix.size());
+        while (!val.empty() && val.front() == ' ') val.remove_prefix(1);
+        resp.content_type = std::string{val};
+    } else if (lower_match(cl_prefix)) {
+        auto val = line.substr(cl_prefix.size());
+        while (!val.empty() && val.front() == ' ') val.remove_prefix(1);
+        std::size_t len = 0;
+        for (char c : val) {
+            if (c < '0' || c > '9') break;
+            len = len * 10 + static_cast<std::size_t>(c - '0');
+        }
+        resp.content_length = len;
+    }
+
+    return total;
+}
+
+std::string resolve_url(std::string_view url) {
+    if (is_s3_url(url))
+        return s3_to_https(url);
+    return std::string{url};
+}
+
+// Per-request auth state returned by apply_auth.
+// Must outlive the curl_easy_perform() call.
+struct AuthState {
+    struct curl_slist* headers = nullptr;
+    std::string sigv4_str;
+    std::string userpwd;
+
+    ~AuthState() {
+        if (headers)
+            curl_slist_free_all(headers);
+    }
+    AuthState() = default;
+    AuthState(const AuthState&) = delete;
+    AuthState& operator=(const AuthState&) = delete;
+    AuthState(AuthState&& o) noexcept
+        : headers(o.headers), sigv4_str(std::move(o.sigv4_str)), userpwd(std::move(o.userpwd))
+    { o.headers = nullptr; }
+    AuthState& operator=(AuthState&& o) noexcept {
+        if (this != &o) {
+            if (headers) curl_slist_free_all(headers);
+            headers = o.headers; o.headers = nullptr;
+            sigv4_str = std::move(o.sigv4_str);
+            userpwd = std::move(o.userpwd);
+        }
+        return *this;
+    }
+};
+
+AuthState apply_auth(CURL* curl, const HttpClient::Config& config) {
+    AuthState state;
+    if (!config.aws_auth.empty()) {
+        // AWS SigV4 takes precedence
+        state.sigv4_str = "aws:amz:" + config.aws_auth.region + ":s3";
+        curl_easy_setopt(curl, CURLOPT_AWS_SIGV4, state.sigv4_str.c_str());
+        state.userpwd = config.aws_auth.access_key + ":" + config.aws_auth.secret_key;
+        curl_easy_setopt(curl, CURLOPT_USERPWD, state.userpwd.c_str());
+        if (!config.aws_auth.session_token.empty()) {
+            auto hdr = "x-amz-security-token: " + config.aws_auth.session_token;
+            state.headers = curl_slist_append(state.headers, hdr.c_str());
+            curl_easy_setopt(curl, CURLOPT_HTTPHEADER, state.headers);
+        }
+    } else if (!config.auth.bearer_token.empty()) {
+        auto header = "Authorization: Bearer " + config.auth.bearer_token;
+        state.headers = curl_slist_append(state.headers, header.c_str());
+        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, state.headers);
+    } else if (!config.auth.username.empty()) {
+        state.userpwd = config.auth.username + ":" + config.auth.password;
+        curl_easy_setopt(curl, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
+        curl_easy_setopt(curl, CURLOPT_USERPWD, state.userpwd.c_str());
+    }
+    return state;
+}
+
+// Read callback for PUT uploads
+struct PutState {
+    std::span<const std::byte> data;
+    std::size_t offset = 0;
+};
+
+std::size_t read_callback(char* buffer, std::size_t size,
+                          std::size_t nmemb, void* userdata) noexcept {
+    auto& state = *static_cast<PutState*>(userdata);
+    auto remaining = state.data.size() - state.offset;
+    auto to_copy = std::min(size * nmemb, remaining);
+    std::memcpy(buffer, state.data.data() + state.offset, to_copy);
+    state.offset += to_copy;
+    return to_copy;
+}
+
+enum class Method { GET, GET_RANGE, HEAD, PUT };
+
+HttpResponse perform(const HttpClient::Config& config,
+                     std::string_view url,
+                     Method method,
+                     std::span<const std::byte> put_data,
+                     std::string_view content_type,
+                     std::string range = {}) {
+    auto resolved = resolve_url(url);
+    HttpResponse resp;
+
+    for (std::size_t attempt = 0; attempt <= config.max_retries; ++attempt) {
+        if (HttpClient::isAborted()) return resp;
+        resp = HttpResponse{};
+        auto* curl = thread_handle();
+        curl_easy_reset(curl);
+
+        // URL
+        curl_easy_setopt(curl, CURLOPT_URL, resolved.c_str());
+
+        // Timeouts
+        curl_easy_setopt(curl, CURLOPT_CONNECTTIMEOUT,
+                         static_cast<long>(config.connect_timeout.count()));
+        curl_easy_setopt(curl, CURLOPT_TIMEOUT,
+                         static_cast<long>(config.transfer_timeout.count()));
+
+        // TCP keepalive: keeps idle-pooled connections from being torn
+        // down between bursts of S3 fetches so we reuse the TLS session.
+        curl_easy_setopt(curl, CURLOPT_TCP_KEEPALIVE, 1L);
+        curl_easy_setopt(curl, CURLOPT_TCP_KEEPIDLE, 30L);
+        curl_easy_setopt(curl, CURLOPT_TCP_KEEPINTVL, 30L);
+
+        // Redirects
+        if (config.follow_redirects) {
+            curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
+            curl_easy_setopt(curl, CURLOPT_MAXREDIRS, 10L);
+        }
+
+        // User agent
+        curl_easy_setopt(curl, CURLOPT_USERAGENT, config.user_agent.c_str());
+
+        // Process-wide abort hook — returning non-zero from the xfer
+        // callback aborts the transfer immediately. Used for fast
+        // shutdown so the worker pool isn't stuck inside curl.
+        curl_easy_setopt(curl, CURLOPT_NOPROGRESS, 0L);
+        curl_easy_setopt(curl, CURLOPT_XFERINFOFUNCTION, xferinfo_callback);
+
+        // Write callback (body)
+        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
+        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &resp.body);
+
+        // Header callback
+        curl_easy_setopt(curl, CURLOPT_HEADERFUNCTION, header_callback);
+        curl_easy_setopt(curl, CURLOPT_HEADERDATA, &resp);
+
+        // Auth (local per-request state)
+        auto auth_state = apply_auth(curl, config);
+
+        // Method-specific setup
+        struct curl_slist* extra_headers = nullptr;
+        PutState put_state;
+        std::string ct_header;  // Must outlive curl_easy_perform
+
+        switch (method) {
+            case Method::GET:
+                curl_easy_setopt(curl, CURLOPT_HTTPGET, 1L);
+                break;
+            case Method::GET_RANGE:
+                curl_easy_setopt(curl, CURLOPT_HTTPGET, 1L);
+                curl_easy_setopt(curl, CURLOPT_RANGE, range.c_str());
+                break;
+            case Method::HEAD:
+                curl_easy_setopt(curl, CURLOPT_NOBODY, 1L);
+                break;
+            case Method::PUT: {
+                curl_easy_setopt(curl, CURLOPT_UPLOAD, 1L);
+                curl_easy_setopt(curl, CURLOPT_INFILESIZE_LARGE,
+                                 static_cast<curl_off_t>(put_data.size()));
+
+                // Set up read data for PUT
+                put_state = PutState{put_data, 0};
+                curl_easy_setopt(curl, CURLOPT_READFUNCTION, read_callback);
+                curl_easy_setopt(curl, CURLOPT_READDATA, &put_state);
+
+                if (!content_type.empty()) {
+                    ct_header = "Content-Type: " + std::string{content_type};
+                    extra_headers = curl_slist_append(extra_headers, ct_header.c_str());
+                    // Merge with auth headers if present
+                    if (auth_state.headers) {
+                        for (auto* node = auth_state.headers; node; node = node->next)
+                            extra_headers = curl_slist_append(extra_headers, node->data);
+                        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, extra_headers);
+                        // Auth headers are copied into extra_headers; free originals
+                        curl_slist_free_all(auth_state.headers);
+                        auth_state.headers = nullptr;
+                    } else {
+                        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, extra_headers);
+                    }
+                }
+                break;
+            }
+        }
+
+        auto code = curl_easy_perform(curl);
+
+        // Clean up request-scoped slist
+        if (extra_headers)
+            curl_slist_free_all(extra_headers);
+        // auth_state cleans up its own headers in destructor
+
+        if (code == CURLE_OK) {
+            curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &resp.status_code);
+
+            // Retry on 5xx server errors
+            if (resp.status_code >= 500 && attempt < config.max_retries) {
+                thread_local std::mt19937 rng{std::random_device{}()};
+                std::uniform_int_distribution<unsigned> jitter(0, 100);
+                std::this_thread::sleep_for(
+                    std::chrono::milliseconds(200 * (1u << attempt) + jitter(rng)));
+                continue;
+            }
+            return resp;
+        }
+
+        // Retry on network / transient curl errors
+        if (attempt < config.max_retries && !HttpClient::isAborted()) {
+            thread_local std::mt19937 rng{std::random_device{}()};
+            std::uniform_int_distribution<unsigned> jitter(0, 100);
+            std::this_thread::sleep_for(
+                std::chrono::milliseconds(200 * (1u << attempt) + jitter(rng)));
+            continue;
+        }
+        break;
+    }
+
+    return resp; // return last response even on failure
+}
+
+} // namespace
+
+// ---------------------------------------------------------------------------
+// HttpClient public API
+// ---------------------------------------------------------------------------
+
+HttpResponse HttpClient::get(std::string_view url) const {
+    return perform(config_, url, Method::GET, {}, {});
+}
+
+HttpResponse HttpClient::get_range(std::string_view url,
+                                   std::size_t offset, std::size_t length) const {
+    if (length == 0) {
+        return HttpResponse{};
+    }
+    auto range = std::to_string(offset) + "-" + std::to_string(offset + length - 1);
+    return perform(config_, url, Method::GET_RANGE, {}, {}, range);
+}
+
+HttpResponse HttpClient::head(std::string_view url) const {
+    return perform(config_, url, Method::HEAD, {}, {});
+}
+
+HttpResponse HttpClient::put(std::string_view url,
+                             std::span<const std::byte> data,
+                             std::string_view content_type) const {
+    return perform(config_, url, Method::PUT, data, content_type);
+}
+
+HttpResponse HttpClient::put_file(std::string_view url,
+                                  const std::filesystem::path& file_path,
+                                  std::string_view content_type) const {
+    auto resolved = resolve_url(url);
+
+    FILE* f = std::fopen(file_path.c_str(), "rb");
+    if (!f) throw std::runtime_error("put_file: cannot open " + file_path.string());
+    std::fseek(f, 0, SEEK_END);
+    auto file_size = std::ftell(f);
+    std::fseek(f, 0, SEEK_SET);
+
+    HttpResponse resp;
+    for (std::size_t attempt = 0; attempt <= config_.max_retries; ++attempt) {
+        if (isAborted()) { std::fclose(f); return resp; }
+        resp = HttpResponse{};
+        std::fseek(f, 0, SEEK_SET);
+        auto* curl = thread_handle();
+        curl_easy_reset(curl);
+
+        curl_easy_setopt(curl, CURLOPT_URL, resolved.c_str());
+        curl_easy_setopt(curl, CURLOPT_CONNECTTIMEOUT,
+                         static_cast<long>(config_.connect_timeout.count()));
+        curl_easy_setopt(curl, CURLOPT_TIMEOUT,
+                         static_cast<long>(config_.transfer_timeout.count()));
+        curl_easy_setopt(curl, CURLOPT_USERAGENT, config_.user_agent.c_str());
+        // Keep TCP connections alive so back-to-back fetches against
+        // S3 don't pay the TLS handshake cost each time. curl_easy_reset
+        // preserves the connection cache and SSL session IDs, so we just
+        // need to ensure keepalive is on. Ping every 30s after 30s idle.
+        curl_easy_setopt(curl, CURLOPT_TCP_KEEPALIVE, 1L);
+        curl_easy_setopt(curl, CURLOPT_TCP_KEEPIDLE, 30L);
+        curl_easy_setopt(curl, CURLOPT_TCP_KEEPINTVL, 30L);
+        if (config_.follow_redirects) {
+            curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
+            curl_easy_setopt(curl, CURLOPT_MAXREDIRS, 10L);
+        }
+        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
+        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &resp.body);
+        curl_easy_setopt(curl, CURLOPT_HEADERFUNCTION, header_callback);
+        curl_easy_setopt(curl, CURLOPT_HEADERDATA, &resp);
+        curl_easy_setopt(curl, CURLOPT_NOPROGRESS, 0L);
+        curl_easy_setopt(curl, CURLOPT_XFERINFOFUNCTION, xferinfo_callback);
+
+        auto auth_state = apply_auth(curl, config_);
+
+        curl_easy_setopt(curl, CURLOPT_UPLOAD, 1L);
+        curl_easy_setopt(curl, CURLOPT_INFILESIZE_LARGE, static_cast<curl_off_t>(file_size));
+        curl_easy_setopt(curl, CURLOPT_READDATA, f);
+        // Default read function is fread — no CURLOPT_READFUNCTION needed
+
+        struct curl_slist* extra_headers = nullptr;
+        std::string ct_header;
+        if (!content_type.empty()) {
+            ct_header = "Content-Type: " + std::string{content_type};
+            extra_headers = curl_slist_append(extra_headers, ct_header.c_str());
+            if (auth_state.headers) {
+                for (auto* node = auth_state.headers; node; node = node->next)
+                    extra_headers = curl_slist_append(extra_headers, node->data);
+                curl_slist_free_all(auth_state.headers);
+                auth_state.headers = nullptr;
+            }
+            curl_easy_setopt(curl, CURLOPT_HTTPHEADER, extra_headers);
+        }
+
+        auto code = curl_easy_perform(curl);
+        if (extra_headers) curl_slist_free_all(extra_headers);
+
+        if (code == CURLE_OK) {
+            curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &resp.status_code);
+            if (resp.status_code >= 500 && attempt < config_.max_retries) {
+                thread_local std::mt19937 rng{std::random_device{}()};
+                std::uniform_int_distribution<unsigned> jitter(0, 100);
+                std::this_thread::sleep_for(
+                    std::chrono::milliseconds(200 * (1u << attempt) + jitter(rng)));
+                continue;
+            }
+            std::fclose(f);
+            return resp;
+        }
+        if (attempt < config_.max_retries) {
+            thread_local std::mt19937 rng{std::random_device{}()};
+            std::uniform_int_distribution<unsigned> jitter(0, 100);
+            std::this_thread::sleep_for(
+                std::chrono::milliseconds(200 * (1u << attempt) + jitter(rng)));
+            continue;
+        }
+    }
+    std::fclose(f);
+    return resp;
+}
+
+} // namespace utils
+
+#endif // UTILS_HAS_CURL
